MSVC and Windows Clang working
It was necessary to temporarily disable address sanitizer in "Debug Developer" mode for msvc.
It was also necessary to use uint32_t instead of uint16_t to please the implicit conversion warnings of msvc.
And it was necessary to reduce the cmake minumum version to 3.22 despite the developer console stating 3.23
And fixed some missing headers in PlaceBuilder.cpp